+++ src/main/java/coordinates/domain/point/Point.java
@@ -1,4 +1,4 @@
-package coordinates.domain;
+package coordinates.domain.point;
 
 import java.util.Objects;
