+++ src/main/java/coordinates/domain/point/XCoordinate.java
@@ -1,4 +1,4 @@
-package coordinates.domain;
+package coordinates.domain.point;
 
 import coordinates.exception.OutOfXCoordinateBoundException;
